+++ SKPhotoBrowser/SKPhotoBrowserOptions.swift
@@ -40,15 +40,15 @@ public struct SKPhotoBrowserOptions {
     public static var swapCloseAndDeleteButtons: Bool = false
     public static var disableVerticalSwipe: Bool = false
 
-    /// Offset from top and from nearest screen edge of close button and delete button.
-    ///
-    /// - Default: 5
-    public static var closeAndDeleteButtonPadding: CGFloat = 5
-    
     /// if this value is true, the long photo width will match the screen,
     /// and the minScale is 1.0, the maxScale is 2.5
     /// Default: false
     public static var longPhotoWidthMatchScreen: Bool = false
+}
+
+public struct SKButtonOptions {
+    public static var closeButtonPadding: CGPoint = CGPoint(x: 5, y: 20)
+    public static var deleteButtonPadding: CGPoint = CGPoint(x: 5, y: 20)
 }
 
 public struct SKCaptionOptions {